ふふふ

+++ UI.cpp
@@ -228,7 +228,7 @@ void UI::EaseScale() {
 	//strong *= 10;
 	strong = player->GetArmWeight()*5;
 	power.clear();
-	for (int tmp = strong; tmp > 0;) {
+	for (int tmp = (int)strong; tmp > 0;) {
 		power.push_back(tmp % 10);
 		tmp /= 10;
 	}
